requirements.txt
langchain==0.1.12
langchain-anthropic==0.1.4
langchain-community==0.0.27
langgraph==0.0.15
chromadb==0.4.24
python-dotenv==1.0.1
tavily-python==0.3.1 